apps/streamlit_app/requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
openpyxl>=3.1.2
pyarrow>=14.0.0
requests>=2.31.0
pyyaml>=6.0.1
# pyspark is available in Databricks runtime; include if testing locally:
# pyspark>=3.5.0
# Databricks Vector Search client (version may vary in your workspace image):
databricks-vectorsearch==0.41.0
databricks-sdk>=0.33.0
